+++ examples/blinky-timer-irq.rs
@@ -14,13 +14,12 @@ use crate::hal::{
     gpio::{gpioa, Output, PushPull},
     pac::{interrupt, Interrupt, Peripherals, TIM2},
     prelude::*,
-    timer::{CountDownTimer, Event, Timer},
+    timer::{CounterUs, Event, Timer},
 };
 
 use core::cell::RefCell;
 use cortex_m::{asm::wfi, interrupt::Mutex};
 use cortex_m_rt::entry;
-use embedded_hal::timer::CountDown;
 
 // NOTE You can uncomment 'hprintln' here and in the code below for a bit more
 // verbosity at runtime, at the cost of throwing off the timing of the blink
@@ -36,14 +35,14 @@ type LedPin = gpioa::PA5<Output<PushPull>>;
 static G_LED: Mutex<RefCell<Option<LedPin>>> = Mutex::new(RefCell::new(None));
 
 // Make timer interrupt registers globally available
-static G_TIM: Mutex<RefCell<Option<CountDownTimer<TIM2>>>> = Mutex::new(RefCell::new(None));
+static G_TIM: Mutex<RefCell<Option<CounterUs<TIM2>>>> = Mutex::new(RefCell::new(None));
 
 // Define an interupt handler, i.e. function to call when interrupt occurs.
 // This specific interrupt will "trip" when the timer TIM2 times out
 #[interrupt]
 fn TIM2() {
     static mut LED: Option<LedPin> = None;
-    static mut TIM: Option<CountDownTimer<TIM2>> = None;
+    static mut TIM: Option<CounterUs<TIM2>> = None;
 
     let led = LED.get_or_insert_with(|| {
         cortex_m::interrupt::free(|cs| {
@@ -79,8 +78,8 @@ fn main() -> ! {
     cortex_m::interrupt::free(|cs| *G_LED.borrow(cs).borrow_mut() = Some(led));
 
     // Set up a timer expiring after 1s
-    let mut timer = Timer::new(dp.TIM2, &clocks).count_down();
-    timer.start(1.hz());
+    let mut timer = Timer::new(dp.TIM2, &clocks).counter();
+    timer.start(1.secs()).unwrap();
 
     // Generate an interrupt when the timer expires
     timer.listen(Event::TimeOut);