+++ examples/timer-syst.rs
@@ -15,7 +15,6 @@ use panic_halt as _;
 use cortex_m_rt::entry;
 use cortex_m_semihosting::hprintln;
 
-use embedded_hal::timer::Cancel;
 use hal::timer;
 use hal::timer::Timer;
 use stm32f4xx_hal as hal;
@@ -30,8 +29,8 @@ fn main() -> ! {
     let clocks = rcc.cfgr.sysclk(24.mhz()).freeze();
 
     // Create a timer based on SysTick
-    let mut timer = Timer::syst(cp.SYST, &clocks).count_down();
-    timer.start(24.hz());
+    let mut timer = Timer::syst(cp.SYST, &clocks).counter();
+    timer.start(42.millis()).unwrap();
 
     hprintln!("hello!").unwrap();
     // wait until timer expires
@@ -47,8 +46,7 @@ fn main() -> ! {
     timer.cancel().unwrap();
 
     // start it again
-    timer.start(24.hz());
-    nb::block!(timer.wait()).unwrap();
+    timer.delay(42.millis()).unwrap();
     hprintln!("timer expired 3").unwrap();
 
     timer.cancel().unwrap();