+++ src/prelude.rs
@@ -57,6 +57,7 @@ pub use embedded_hal::watchdog::WatchdogEnable as _embedded_hal_watchdog_Watchdo
 pub use embedded_hal::Capture as _embedded_hal_Capture;
 pub use embedded_hal::Pwm as _embedded_hal_Pwm;
 pub use embedded_hal::Qei as _embedded_hal_Qei;
+pub use fugit::ExtU32 as _fugit_ExtU32;
 
 #[cfg(all(feature = "device-selected", feature = "dac"))]
 pub use crate::dac::DacExt as _stm32f4xx_hal_dac_DacExt;